[#1540] Replace assertEquals with assertTrue where applicable. Fix checkstyle errors.

+++ gradoop-flink/src/test/java/org/gradoop/flink/io/impl/statistics/GraphStatisticsDataSinkTest.java
@@ -44,7 +44,7 @@ public class GraphStatisticsDataSinkTest extends GradoopFlinkTestBase {
 
     GraphStatistics statistics = GraphStatisticsLocalFSReader.read(tmpPath);
 
-    assertEquals(11L, (long)statistics.getVertexCount());
+    assertEquals(11L, statistics.getVertexCount());
   }
